requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
tavily-python
python-dotenv
requests
jinja2
